test: ChessGameTest 생성
